+++ packages/react/src/experimental/Navigation/ApplicationFrame/index.stories.tsx
@@ -1,15 +1,15 @@
 import type { Meta, StoryObj } from "@storybook/react"
 import { ComponentProps } from "react"
-import { ApplicationFrame } from "./index"
 import { Page } from "../Page"
 import * as PageStories from "../Page/index.stories"
 import * as SidebarStories from "../Sidebar/index.stories"
 import { Sidebar } from "../Sidebar/Sidebar"
+import { ApplicationFrame } from "./index"
 
 const meta: Meta<typeof ApplicationFrame> = {
   title: "ApplicationFrame",
   component: ApplicationFrame,
-  tags: ["autodocs", "experimental", "no-sidebar"],
+  tags: ["autodocs", "experimental"],
   parameters: {
     layout: "fullscreen",
   },