+++ packages/react/src/experimental/Navigation/Sidebar/Header/index.stories.tsx
@@ -1,5 +1,5 @@
-import { Image } from "../../../../icons/app"
 import type { Meta, StoryObj } from "@storybook/react"
+import { Image } from "../../../../icons/app"
 import { SidebarHeader } from "./index"
 
 const meta = {