+++ packages/react/src/experimental/Navigation/Sidebar/Menu/index.stories.tsx
@@ -1,5 +1,5 @@
-import * as Icons from "../../../../icons/app"
 import type { Meta, StoryObj } from "@storybook/react"
+import * as Icons from "../../../../icons/app"
 import { Menu } from "./index"
 
 const meta = {
@@ -25,6 +25,7 @@ export const Default: Story = {
   args: {
     tree: [
       {
+        id: "1",
         title: "Root",
         items: [
           { label: "Home", icon: Icons.Home, href: "/", exactMatch: true },
@@ -41,8 +42,10 @@ export const Default: Story = {
           },
         ],
         isRoot: true,
+        isSortable: false,
       },
       {
+        id: "2",
         title: "You",
         items: [
           { label: "Me", icon: Icons.Person, href: "/me" },
@@ -50,15 +53,26 @@ export const Default: Story = {
           { label: "Time off", icon: Icons.PalmTree, href: "/time-off" },
         ],
         isOpen: true,
+        isSortable: true,
       },
       {
+        id: "3",
         title: "Your company",
         items: [
           { label: "Organization", icon: Icons.People, href: "/organization" },
           { label: "Calendar", icon: Icons.Calendar, href: "/calendar" },
         ],
         isOpen: true,
+        isSortable: true,
       },
     ],
+    onCollapse: (category, isOpen) => {
+      console.log(
+        `Category "${category.title}" is now ${isOpen ? "open" : "closed"}`
+      )
+    },
+    onSort: (categories) => {
+      console.log("Categories sorted:", categories)
+    },
   },
 }