+++ packages/react/src/ui/chart.tsx
@@ -2,8 +2,8 @@ import * as React from "react"
 import { useLayoutEffect, useMemo, useState } from "react"
 import * as RechartsPrimitive from "recharts"
 
-import { cn } from "../lib/utils"
 import { cva, type VariantProps } from "cva"
+import { cn } from "../lib/utils"
 
 const variants = cva({
   variants: {